+++ sciencebeam_judge/evaluation/custom/bounding_box_intersection.py
